+++ Phase_3_tests/Android_Tests/readme.md
@@ -1,0 +1,3 @@
+# The Android Application Tests Folder
+
+## E2E Tests that covers 95% of the functionalities
