+++ Phase_3_tests/Website_Tests/readme.md
@@ -1,0 +1,3 @@
+# The website Pages Tests Folder
+
+## E2E Tests that covers 95% of the functionalities
