--- groups.test.js
@@ -1,26 +1,0 @@
-/// <reference types="cypress" />
-
-let password = Cypress.env("password")
-let email = Cypress.env("email")
-
-describe("Testing Albums", () => {
-    before("Pre-test: Login...", () => {
-        cy.viewport(1366, 768)
-        cy.visit("https://identity.flickr.com/account-picker")
-
-        cy.contains("Use another account", { timeout: 10000 }).click()
-        cy.get("input[type=email").type("mohamed84348@gmail.com")
-        cy.get("[data-testid=identity-form-submit-button]").click()
-
-        cy.get("input[type=password]", { timeout: 10000 }).type("R/FUdqJn?8hd+L+")
-        cy.get("[data-testid=identity-form-submit-button]").click()
-        cy.get(".you").click()
-
-    })
-
-
-    it("Groups Test...", () => {
-        cy.get("#groups").click()
-        cy.contains("Flikr_Test_Group").should("exist")
-    })
-})
